jira-kpi-to-sheets/requirements.txt
requests
python-dotenv
gspread
os
datetime
dateutil.parser